requirements.txt
decord 
pillow 
sentencepiece
opencv-python-headless 
pytesseract
hf_xet
moviepy
langchain 
langchain-core 
langchain_community 
langchain_huggingface
faster-whisper
streamlit